Move supporting functions to helper.py
  Update Tests/kaas/k8s-default-storage-class/k8s-default-storage-class-check.py
  Update Tests/kaas/k8s-default-storage-class/helper.py
